requirements.txt
PyGithub==2.1.1
python-dotenv==1.0.0
PyYAML==6.0.1
click==8.1.7
inquirer==3.1.3

